Split document_helpers into existing owner modules
Move Writer text, document resolution, streamed-edit, Calc chat
context, paragraph-range, and heading-bookmark helpers out of the
grab-bag so LibrePy can import text_helpers without loading
SheetAnalyzer. document_helpers keeps the Writer chat-context
assembler, Draw chat context, resolve_locator, and DocumentService.
