+++ docs/repo-map.md
@@ -25,8 +25,8 @@ Start here by task.
 | Agent backends | Optional external backends (`agent_backend.backend_id` when not `builtin`) | [`plugin/agent_backend/`](../plugin/agent_backend/) |
 | HTTP / LLM | Chat requests, tools, token stripping, pacing | [`plugin/framework/client/llm_client.py`](../plugin/framework/client/llm_client.py) (`make_chat_request`, `request_with_tools`, …), [`plugin/framework/client/http_transport.py`](../plugin/framework/client/http_transport.py), [`plugin/framework/client/auth.py`](../plugin/framework/client/auth.py) |
 | Tools registry | Tool registration and schemas | [`plugin/framework/tool.py`](../plugin/framework/tool.py) |
-| UNO document helpers | Chat context, undo/stream, URL/frame resolve (WriterAgent) | [`plugin/doc/document_helpers.py`](../plugin/doc/document_helpers.py) |
-| Light document helpers | Linebreaks, tracked-deletion reads, heading tree, path; type guards; UD props (LibrePy-safe) | [`plugin/doc/text_helpers.py`](../plugin/doc/text_helpers.py), [`plugin/doc/doc_type.py`](../plugin/doc/doc_type.py), [`plugin/doc/udprops.py`](../plugin/doc/udprops.py) |
+| UNO document helpers | Writer chat-context assembler and `DocumentService` (WriterAgent) | [`plugin/doc/document_helpers.py`](../plugin/doc/document_helpers.py) |
+| Light document helpers | Linebreaks, tracked-deletion reads, heading tree, path, selection text, Writer text slices; type guards; UD props (LibrePy-safe) | [`plugin/doc/text_helpers.py`](../plugin/doc/text_helpers.py), [`plugin/doc/doc_type.py`](../plugin/doc/doc_type.py), [`plugin/doc/udprops.py`](../plugin/doc/udprops.py) |
 | Config / keys / LRU | `writeragent.json`, keys, LRU | [`plugin/framework/config.py`](../plugin/framework/config.py) |
 | Dialogs / XDL | Dialog load helpers and settings UI | [`plugin/chatbot/dialogs.py`](../plugin/chatbot/dialogs.py), [`plugin/chatbot/dialog_views.py`](../plugin/chatbot/dialog_views.py), [`plugin/chatbot/settings_dialog.py`](../plugin/chatbot/settings_dialog.py) |
 | Async UI drain | Stream queue drain on the UI thread (`get_toolkit`, `get_ctx`) | [`plugin/framework/async_stream.py`](../plugin/framework/async_stream.py), [`plugin/framework/uno_context.py`](../plugin/framework/uno_context.py) |
