+++ tests/writer/test_document_helpers_uno.py
@@ -1,6 +1,10 @@
 
 from plugin.doc.text_helpers import build_heading_tree, get_string_without_tracked_deletions
-from plugin.doc.document_helpers import resolve_locator, get_paragraph_ranges, get_document_length, ensure_heading_bookmarks, get_document_context_for_chat, WriterStreamedRewriteSession
+from plugin.doc.document_helpers import resolve_locator, get_document_context_for_chat
+from plugin.doc.paragraph_search import get_paragraph_ranges
+from plugin.doc.text_helpers import get_document_length
+from plugin.writer.edit_review import WriterStreamedRewriteSession
+from plugin.writer.specialized.bookmarks import ensure_heading_bookmarks
 from plugin.testing_runner import native_test
 from plugin.tests.testing_utils import TestingFactory, with_native_doc
