+++ tests/writer/test_track_changes_reviewable_uno.py
@@ -17,7 +17,7 @@ import contextlib
 
 from plugin.testing_runner import native_test
 from plugin.tests.testing_utils import TestingFactory, with_native_doc
-from plugin.doc.document_helpers import WriterStreamedRewriteSession, WriterStreamedAppendSession
+from plugin.writer.edit_review import WriterStreamedRewriteSession, WriterStreamedAppendSession
 from plugin.writer.content import ApplyDocumentContent
 import plugin.writer.edit_review as _content
 from plugin.writer.edit_review import EditReviewSession, get_agent_edit_review_mode
